Move quaternion math functions from toolbox to coordinates
 * Minor change to quaternionNormalize implementation to avoid bringing
   in toolbox.
