+++ Doc/source/coordinates.rst
@@ -8,8 +8,23 @@ coordinates - module for coordinate transforms
 
 .. currentmodule:: spacepy.coordinates
 
+.. rubric:: Classes
+
 .. autosummary::
     :template: clean_class.rst
     :toctree: autosummary
 
     Coords
+
+.. rubric:: Functions
+
+.. autosummary::
+   :template: clean_function.rst
+   :toctree: autosummary
+
+    quaternionRotateVector
+    quaternionNormalize
+    quaternionMultiply
+    quaternionConjugate
+    quaternionFromMatrix
+    quaternionToMatrix
